+++ next-auth.d.ts
@@ -1,13 +1,22 @@
-import { Role } from "@prisma/client";
 import NextAuth, { type DefaultSession } from "next-auth";
+import { Role } from "@prisma/client";
 
 export type ExtendedUser = DefaultSession["user"] & {
-  role: Role;
-  isOAuth: boolean;
+    role: Role;
 };
 
 declare module "next-auth" {
-  interface Session {
-    user: ExtendedUser;
-  }
+    interface Session {
+        user: ExtendedUser;
+        error?: "RefreshTokenError"
+    }
+}
+   
+declare module "next-auth/jwt" {
+    interface JWT {
+      access_token: string
+      expires_at: number
+      refresh_token?: string
+      error?: "RefreshTokenError"
+    }
 }
